move DisplayProtocolName to apply.rs and make pub(crate)
