+++ core/src/upgrade/apply.rs
@@ -18,9 +18,7 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::upgrade::{
-    DisplayProtocolName, InboundUpgrade, OutboundUpgrade, ProtocolName, UpgradeError,
-};
+use crate::upgrade::{InboundUpgrade, OutboundUpgrade, ProtocolName, UpgradeError};
 use crate::{connection::ConnectedPoint, Negotiated};
 use futures::{future::Either, prelude::*};
 use log::debug;
@@ -29,6 +27,7 @@ use std::{iter, mem, pin::Pin, task::Context, task::Poll};
 
 pub use multistream_select::Version;
 use smallvec::SmallVec;
+use std::fmt;
 
 // TODO: Still needed?
 /// Applies an upgrade to the inbound and outbound direction of a connection or substream.
@@ -272,5 +271,41 @@ struct NameWrap<N>(N);
 impl<N: ProtocolName> AsRef<[u8]> for NameWrap<N> {
     fn as_ref(&self) -> &[u8] {
         self.0.protocol_name()
+    }
+}
+
+/// Wrapper for printing a [`ProtocolName`] that is expected to be mostly ASCII
+pub(crate) struct DisplayProtocolName<N>(pub N);
+
+impl<N: ProtocolName> fmt::Display for DisplayProtocolName<N> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        use fmt::Write;
+        for byte in self.0.protocol_name() {
+            if (b' '..=b'~').contains(byte) {
+                f.write_char(char::from(*byte))?;
+            } else {
+                write!(f, "<{:02X}>", byte)?;
+            }
+        }
+        Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn display_protocol_name() {
+        assert_eq!(DisplayProtocolName(b"/hello/1.0").to_string(), "/hello/1.0");
+        assert_eq!(DisplayProtocolName("/hellö/").to_string(), "/hell<C3><B6>/");
+        assert_eq!(
+            DisplayProtocolName((0u8..=255).collect::<Vec<_>>()).to_string(),
+            (0..32)
+                .map(|c| format!("<{:02X}>", c))
+                .chain((32..127).map(|c| format!("{}", char::from_u32(c).unwrap())))
+                .chain((127..256).map(|c| format!("<{:02X}>", c)))
+                .collect::<String>()
+        );
     }
 }
